backend/requirement.txt
ingredient_parser_nlp==2.1.1
langchain_community==0.3.27
bcrypt==4.0.1
langchain_google_genai==2.1.8
langchain==0.3.27
selenium==4.34.2
jq==1.10.0
chromadb==1.0.15
langchain_chroma==0.2.5
langgraph==0.6.2
fastapi==0.116.1
uvicorn==0.27.1
aiosqlite==0.21.0
python-dotenv==1.1.0
passlib==1.7.4
PyJWT>=2.9.0        
python-multipart==0.0.20